# reimplement wyhash v4.1