This updates to the new v4 version, so existing hashes will fail and this is not backwards compatible.

Be skeptical of the true value below as there are needed improvements to the benchmarking script. Use this as a relative performance indicator.

### Before

```
$ ./benchmark --filter wyhash --key-size 65535
wyhash
   iterative: 15754 MiB/s [ac7dfbc737b3ee33]
  small keys: 14744 MiB/s [c504fcf016bbda43]
```

### After

```
$ ./benchmark --filter wyhash --key-size 65535
wyhash
   iterative: 31050 MiB/s [a2ad8475e77f0a2a]
  small keys: 28081 MiB/s [56f43308c6a8af3f]
```

---

Compared to the C reference we are slightly faster:

Our 16-48 byte keys however have worse performance however, which should be investigated but I wouldn't consider that a blocker.

---

Full measurements from my [smhasher fork](https://github.com/tiehuis/smhasher) below:

```
 (master =) $ ./SMHasher --test=Speed zig-wyhash
--- Testing zig-wyhash "zig wyhash v4.1 (64-bit)" GOOD

[[[ Speed Tests ]]]

Bulk speed test - 262144-byte keys
Alignment  7 -  9.122 bytes/cycle - 26097.00 MiB/sec @ 3 ghz
Alignment  6 -  9.121 bytes/cycle - 26094.49 MiB/sec @ 3 ghz
Alignment  5 -  9.119 bytes/cycle - 26090.35 MiB/sec @ 3 ghz
Alignment  4 -  9.132 bytes/cycle - 26128.15 MiB/sec @ 3 ghz
Alignment  3 -  9.118 bytes/cycle - 26086.76 MiB/sec @ 3 ghz
Alignment  2 -  9.119 bytes/cycle - 26089.19 MiB/sec @ 3 ghz
Alignment  1 -  9.119 bytes/cycle - 26088.81 MiB/sec @ 3 ghz
Alignment  0 -  9.145 bytes/cycle - 26163.94 MiB/sec @ 3 ghz
Average      -  9.124 bytes/cycle - 26104.84 MiB/sec @ 3 ghz

Small key speed test -    1-byte keys -    16.00 cycles/hash
Small key speed test -    2-byte keys -    16.00 cycles/hash
Small key speed test -    3-byte keys -    16.00 cycles/hash
Small key speed test -    4-byte keys -    17.00 cycles/hash
Small key speed test -    5-byte keys -    17.00 cycles/hash
Small key speed test -    6-byte keys -    17.00 cycles/hash
Small key speed test -    7-byte keys -    17.00 cycles/hash
Small key speed test -    8-byte keys -    17.00 cycles/hash
Small key speed test -    9-byte keys -    17.00 cycles/hash
Small key speed test -   10-byte keys -    17.00 cycles/hash
Small key speed test -   11-byte keys -    17.00 cycles/hash
Small key speed test -   12-byte keys -    17.00 cycles/hash
Small key speed test -   13-byte keys -    17.00 cycles/hash
Small key speed test -   14-byte keys -    17.00 cycles/hash
Small key speed test -   15-byte keys -    17.00 cycles/hash
Small key speed test -   16-byte keys -    17.00 cycles/hash
Small key speed test -   17-byte keys -    25.00 cycles/hash
Small key speed test -   18-byte keys -    25.00 cycles/hash
Small key speed test -   19-byte keys -    25.00 cycles/hash
Small key speed test -   20-byte keys -    25.00 cycles/hash
Small key speed test -   21-byte keys -    25.00 cycles/hash
Small key speed test -   22-byte keys -    25.00 cycles/hash
Small key speed test -   23-byte keys -    25.00 cycles/hash
Small key speed test -   24-byte keys -    25.00 cycles/hash
Small key speed test -   25-byte keys -    25.00 cycles/hash
Small key speed test -   26-byte keys -    25.00 cycles/hash
Small key speed test -   27-byte keys -    25.00 cycles/hash
Small key speed test -   28-byte keys -    25.00 cycles/hash
Small key speed test -   29-byte keys -    25.00 cycles/hash
Small key speed test -   30-byte keys -    25.00 cycles/hash
Small key speed test -   31-byte keys -    25.00 cycles/hash
Average                                    20.774 cycles/hash


Input vcode 0x00000001, Output vcode 0x00000001, Result vcode 0x00000001
Verification value is 0x00000001 - Testing took 5.235457 seconds
 (master =) $ ./SMHasher --test=Speed wyhash
--- Testing wyhash "wyhash v4.1 (64-bit)" GOOD

[[[ Speed Tests ]]]

Bulk speed test - 262144-byte keys
Alignment  7 -  8.695 bytes/cycle - 24876.65 MiB/sec @ 3 ghz
Alignment  6 -  8.695 bytes/cycle - 24876.62 MiB/sec @ 3 ghz
Alignment  5 -  8.682 bytes/cycle - 24840.46 MiB/sec @ 3 ghz
Alignment  4 -  8.748 bytes/cycle - 25028.81 MiB/sec @ 3 ghz
Alignment  3 -  8.694 bytes/cycle - 24874.31 MiB/sec @ 3 ghz
Alignment  2 -  8.710 bytes/cycle - 24920.91 MiB/sec @ 3 ghz
Alignment  1 -  8.707 bytes/cycle - 24912.00 MiB/sec @ 3 ghz
Alignment  0 -  8.662 bytes/cycle - 24782.68 MiB/sec @ 3 ghz
Average      -  8.699 bytes/cycle - 24889.05 MiB/sec @ 3 ghz

Small key speed test -    1-byte keys -    17.00 cycles/hash
Small key speed test -    2-byte keys -    17.00 cycles/hash
Small key speed test -    3-byte keys -    17.00 cycles/hash
Small key speed test -    4-byte keys -    17.81 cycles/hash
Small key speed test -    5-byte keys -    17.78 cycles/hash
Small key speed test -    6-byte keys -    17.77 cycles/hash
Small key speed test -    7-byte keys -    17.77 cycles/hash
Small key speed test -    8-byte keys -    17.78 cycles/hash
Small key speed test -    9-byte keys -    17.80 cycles/hash
Small key speed test -   10-byte keys -    17.75 cycles/hash
Small key speed test -   11-byte keys -    17.77 cycles/hash
Small key speed test -   12-byte keys -    17.77 cycles/hash
Small key speed test -   13-byte keys -    17.80 cycles/hash
Small key speed test -   14-byte keys -    17.76 cycles/hash
Small key speed test -   15-byte keys -    17.76 cycles/hash
Small key speed test -   16-byte keys -    17.77 cycles/hash
Small key speed test -   17-byte keys -    21.00 cycles/hash
Small key speed test -   18-byte keys -    20.00 cycles/hash
Small key speed test -   19-byte keys -    20.00 cycles/hash
Small key speed test -   20-byte keys -    20.00 cycles/hash
Small key speed test -   21-byte keys -    20.95 cycles/hash
Small key speed test -   22-byte keys -    20.16 cycles/hash
Small key speed test -   23-byte keys -    20.00 cycles/hash
Small key speed test -   24-byte keys -    20.00 cycles/hash
Small key speed test -   25-byte keys -    20.75 cycles/hash
Small key speed test -   26-byte keys -    20.00 cycles/hash
Small key speed test -   27-byte keys -    20.00 cycles/hash
Small key speed test -   28-byte keys -    20.14 cycles/hash
Small key speed test -   29-byte keys -    20.48 cycles/hash
Small key speed test -   30-byte keys -    20.17 cycles/hash
Small key speed test -   31-byte keys -    20.00 cycles/hash
Average                                    18.895 cycles/hash


Input vcode 0x00000001, Output vcode 0x00000001, Result vcode 0x00000001
Verification value is 0x00000001 - Testing took 4.933275 seconds
```