@tiehuis 2023-06-07
Given the default hash in the compiler is currently wyhash, what are the implications for the version change? Are there any required invalidations (e.g. for caches) that are required?