@andrewrk 2023-06-07
Nice work!

No, there are no required invalidations. Zig already uses its own version string in hashes for cache purposes.