@Jarred-Sumner 2023-06-11
> Given the default hash in the compiler is currently wyhash, what are the implications for the version change? Are there any required invalidations (e.g. for caches) that are required?

Bun currently uses Wyhash for a handful of things which are visible to users. This is a breaking change for Bun, but it is reasonable to suggest we should just copy-paste the old code into our codebase to ensure it doesn't change across Zig versions